[ADD] analytic_partner
Classify analytic items by partner
==================================

This module adds a partner on each analytic item for allowing to have another
dimension for analysing data.

It also handles the proper propagation of this field to the created analytic
entries when validating invoices.

Configuration
=============

You have to be granted as at least "Accountant" in your user profile and
have checked the "Analytic Accounting" access right.

Usage
=====

Go to Accounting > Analytic Journal Items, and there, you can set the partner
for the analytic items, and search or group by it.

You can also go to a partner, and click on the smart-button "Cost/Revenue",
placed on the upper-right part, and you will navigate to the analytic items
associated to this partner.

Known issues / Roadmap
======================

* This module hasn't been tested with *account_analytic_plans* module
installed, so maybe it's incompatible with it.