[IMP] stock_analytic: Take into account extra move lines flow
If we add extra move lines in stock pickings and wanted to set the
analytic account on it, as the related move is not yet created, the
behaviour of the related field is not wanted.

So, store the analytic account on move line level too and add it in
the values preparation hooks.